move context menu handler to separate module
